requirements.txt
# 核心框架 - A2A SDK 和 ADK
# a2a-sdk is the correct package for building servers and clients.
a2a-sdk>=0.1.0
google-adk>=0.1.0

# AI模型库 - Google Gemini
# This library allows our agents to connect to the Gemini model.
google-generativeai>=0.5.0

# MCP 客户端 - 用于与MCP服务器通信
# This library is needed to talk to the GitHub MCP Server.
mcp>=0.0.3

# Web服务器与HTTP客户端
# Uvicorn runs our A2A agent servers, and httpx is used by the client to send requests.
uvicorn>=0.29.0
httpx>=0.27.0

# 辅助工具
# python-dotenv loads configuration from the .env file.
# nest_asyncio allows asyncio to run in environments that already have an event loop.
python-dotenv>=1.0.0
nest_asyncio>=1.6.0